Extract get_workspace_default_members to util
